pyproject.toml
[project]
name = "boem"
description = "BOEM bird detection and classification pipeline"
readme = "README.md"
requires-python = ">=3.10"
version = "0.0.0"
dependencies = [
  "numpy",
  "pandas",
  "scikit-learn",
  "hydra-core",
  "omegaconf",
  "geopandas",
  "rasterio",
  "dask",
  "bokeh",
  "matplotlib",
  "paramiko",
  "torch",
  "torchmetrics",
  "pytorch-lightning",
  "label-studio-sdk",
  "label-studio-converter",
  "comet-ml",
  "deepforest",
  "globus-sdk",
  # H-CAST requirements
  "timm",
  "opencv-contrib-python",
  "deepforest<2.0",
  # DGL requires CUDA-specific wheels - will be resolved from find-links configured below
  "dgl",
  "packaging",
  "pytest",
  "python-dotenv>=1.2.1",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128" }
]
torchvision = [
  { index = "pytorch-cu128" }
]
deepforest = [
  { path = "/home/b.weinstein/DeepForest" }
]

[tool.uv]
# DGL installation: Configure find-links to point to DGL wheel repository
# DGL requires CUDA-specific wheels. Adjust the URL for your PyTorch/CUDA version.
# See https://www.dgl.ai/pages/start.html for available versions
find-links = [
    "https://data.dgl.ai/wheels/torch-2.1/cu121/repo.html"
]
# Exclude
override-dependencies = [
  "opencv-python; python_version < '0'",
  "opencv-python-headless; python_version < '0'",
]

